test on real file
