+++ tests/data/catalog.csv
@@ -1,0 +1,2 @@
+"dataset.id";"dataset.title";"dataset.slug";"dataset.url";"dataset.organization";"dataset.organization_id";"dataset.license";"dataset.private";"id";"url";"title";"description";"filetype";"format";"mime";"filesize";"checksum.type";"checksum.value";"created_at";"modified";"downloads";"harvest.modified_at"
+"601ddcfc85a59c3a45c2435a";"";"";"";"";"";"";"";"c4e3a9fb-4415-488e-ba57-d05269b27adf";"https://example.com/resource-1";"";"";"";"";"";"";"";"";"";"";"";""
